backend/requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
python-dotenv==0.12.0
PyJWT==1.7.1
pymysql==0.9.3
requests==2.23.0
tabulate==0.8.7
